refactor(critical): drop widget-mode flag, ship the new widget UI everywhere (ARC-753)
The `NEXT_PUBLIC_CRITICAL_WIDGET_MODE` flag has gated the new MatchWidget
behind a default-off env var since ARC-631. Production has been serving
the legacy table layout this whole time even though both branches carry
the new widget code — that's why prod (Critical · The Short Circuit) looked
nothing like local. Remove the flag and the legacy renderer so users get
the redesigned UI by default.

Changes:
- Delete `useWidgetMode`, `widgetMode` lib + tests (the flag infra).
- Delete `ActiveGameContent` and the entire legacy component tree it
  owned: `CriticalGameHeader`, `MatchHud`, `GameTableSection`,
  `CenterTableSection`, `TablePlayer`, `PlayerHand`, `HandControlsPanel`,
  `HandLayoutDropdown`, `ActionsSection`, `ExpansionActions`,
  `MobileHandPopover`, `CardActionsPopover`.
- Port the autoplay menu + idle-timer countdown that lived in the
  legacy `CriticalGameHeader` into a top bar inside `MatchWidget` so
  players still get manual autoplay control and the visible countdown.
- Drop now-dead `actionBusy`, `currentTurnPlayer`, `handLayout`,
  `setHandLayout`, `handleUnstash`, `handleOpenFavorModal`,
  `HandLayoutMode`, and `HAND_LAYOUT_OPTIONS`.
- Inline `MatchWidgetProps` (was `ComponentProps<typeof ActiveGameContent>`)
  so the widget no longer depends on the deleted legacy renderer.
- Extract `readHandToggle` / `writeHandToggle` into
  `lib/handToggleStorage.ts` to keep `MatchWidget.tsx` under the 500-line
  cap after the new top-bar additions.
