+++ apps/web/src/widgets/CriticalGame/lib/constants/ui.ts
@@ -1,12 +1,1 @@
-import { HandLayoutMode } from '../../types';
-
-export const HAND_LAYOUT_OPTIONS: HandLayoutMode[] = [
-  'grid',
-  'grid-3',
-  'grid-4',
-  'grid-5',
-  'grid-6',
-  'linear',
-];
-
 export const SPRITE_GRID_SIZE = 7; // 7x7 grid
