+++ apps/web/src/widgets/CriticalGame/ui/MatchWidget.test-fixtures.ts
@@ -34,10 +34,6 @@ export function makeProps(
     isMyTurn: true,
     canAct: true,
     canPlayNope: false,
-    actionBusy: null,
-    aliveOpponents: [],
-    handLayout: 'grid',
-    setHandLayout: vi.fn(),
     resolveDisplayName: (_id?: string, fb?: string) => fb,
     t: ((k: string) => k) as never,
     actions: {
@@ -49,16 +45,17 @@ export function makeProps(
       playDefuse: vi.fn(),
       playActionCard: vi.fn(),
     } as never,
-    idleTimerTriggered: false,
-    autoplayState: { setAllEnabled: vi.fn() },
-    handleUnstash: vi.fn(),
     handlePlayActionCard: vi.fn(),
-    handleOpenFavorModal: vi.fn(),
     handleOpenEventCombo: vi.fn(),
     handleOpenFiverCombo: vi.fn(),
     formatLogMessage: (m?: string | null) => m ?? '',
     isFullscreen: false,
     toggleFullscreen: vi.fn(),
+    autoplayState: { setAllEnabled: vi.fn() } as never,
+    idleTimerEnabled: false,
+    idleTimerTriggered: false,
+    handleIdleTimeout: vi.fn(),
+    handleStopAutoplay: vi.fn(),
     ...override,
   } as MatchWidgetProps;
 }
